fix(config): reject over-int connect_timeout instead of wrapping through (int) cast
Both ws config paths read connect_timeout via view.getLong and cast to
int: connect_timeout=4294967297 silently became a 1 ms timeout, and
2147483648 wrapped to Integer.MIN_VALUE, producing the misleading
'connect_timeout must be > 0: -2147483648'.

- ConfigView.getInt now parses in the long domain: schema range still
  reports first (tighter, key-specific), then an over-int value fails
  with the actual limit ('connect_timeout must be <= 2147483647: ...')
  instead of a generic 'invalid' or a wrapped value.
- Sender.fromConfigWebSocket and QwpQueryClient.fromConfig/validateConfig
  read connect_timeout via getInt, aligning with the legacy http/tcp
  parseIntValue path that already rejected over-int values.

Covered at ConfigView level (ConfigViewTest) and end-to-end on both
facades (QwpConfigKeysTest.testConnectTimeoutOverIntRejectedOnBoth).

+++ core/src/main/java/io/questdb/client/Sender.java
@@ -3599,7 +3599,9 @@ public interface Sender extends Closeable, ArraySender<Sender> {
                     authTimeoutMillis(view.getLong("auth_timeout_ms", 0));
                 }
                 if (view.has("connect_timeout")) {
-                    connectTimeoutMillis((int) view.getLong("connect_timeout", 0));
+                    // getInt (not getLong + cast): connectTimeoutMillis takes an
+                    // int, and an over-int value must reject, not wrap.
+                    connectTimeoutMillis(view.getInt("connect_timeout", 0));
                 }
 
                 s = view.getStr("auto_flush_rows");
